Add two-way relationship and http response codes

+++ emp/src/main/java/com/example/emp/controller/CompanyController.java
@@ -4,7 +4,9 @@ import com.example.emp.model.Company;
 import com.example.emp.model.CompanyInfo;
 import com.example.emp.model.dto.CompanyDto;
 import com.example.emp.service.CompanyService;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import java.util.List;
 
 @RestController
@@ -28,6 +30,10 @@ public class CompanyController
     @GetMapping(value = "/{company_id}")
     public Company getCompany(@PathVariable long company_id)
     {
+        if(!companyService.companyExists(company_id))
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Company with ID: " + company_id + " does not exist");
+        }
         return companyService.getCompany(company_id);
     }
 
@@ -40,12 +46,20 @@ public class CompanyController
     @PutMapping(value = "/{company_id}")
     public Company modifyCompany(@PathVariable long company_id, @RequestBody CompanyDto companyDto)
     {
+        if(!companyService.companyExists(company_id))
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Company with ID: " + company_id + " does not exist");
+        }
         return companyService.modifyCompany(company_id, companyDto);
     }
 
     @DeleteMapping(value = "/{company_id}")
     public Company deleteCompany(@PathVariable long company_id)
     {
+        if(!companyService.companyExists(company_id))
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Company with ID: " + company_id + " does not exist");
+        }
         return companyService.deleteCompany(company_id);
     }
 }
